backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
PyJWT==2.6.0
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3
pymysql==1.0.3
cryptography==40.0.1
